Prune the deprecated/discouraged parts of the API
The old `StatusError` and related types were formally deprecated a while
back. There should be roughly zero actively maintained code still using
it, so now is as good a time as any to rip off that bandage.

While `StackErr` was not formally deprecated, the documented best
practices suggest that the type itself should never be directly
referenced. Since that's the case, we shift to a unexported type and
remove that as well.

Finally, and this is something that could surely be up for debate, I
found the default stack trace format to be incredible dense and hard to
read, so I added one that resembles the output format of `panic` and
matches zap's default stack tracing format.

Also swap golangci-lint to a YAML config, because the TOML config
doesn't seem to be parsing anymore.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>