feat!: remove custom error Format function (#12)
Using print formatting verbs to opt into stack trace is a smart concept,
but in practice it falls short for a few reasons.

The first is that in real world scenarios, the selection of formatting
verb is not done specifically with error stack traces in mind. This
disconnect is most obvious in slog's handling of "any" types, including
errors. When slog serializes an "any" type, the `%+v` verb is used, and
stack traces are automatically included. Unfortunately this means that
by default, the default slog logger and the text logger include the full
stack trace as an inline value in a logfmt-style key/value pair. This is
virtually never what someone would want.

The good news is, when someone explicitly wants a stack trace, there is
already a mechanism by which they can get that, which is by using
`errors.As` with `serrors.StackTracer`. Even though the level of effort
to opt-in is a bit higher, it means slog handlers can be written to
include stack traces without worrying about making the default logging
behavior harder to parse.

The second is that it plays poorly with error wrapping. The original
xerrors design had iteration in mind--each error after formatting itself
would return the next error in the error chain. If this got adoption in
the standard library, it would probably be a safe assumption that any
error wrapper would be able to delegate formatting appropriately down
the chain, and any error wrapper not conforming to this pattern was out
of compliance, the same way that `Unwrap` support is assumed today.

Unfortunately, serrors was not originally designed to do this, and even
if it were, it requires every wrapper up the stack to adhere to the same
principle. Since a simple `fmt.Errorf` will swallow any custom
formatting, the error printing behavior becomes unreliable based on what
wrappers are in use.

Because this removes one of the three documented ways to get the stack
trace out of an error, it is considered a breaking change. I would not
expect it to break applications, but the change in behavior may be
meaningful for application logs where stack traces may need to be
explicitly opted into using one of the other two documented methods.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>