ci: bump tool versions
Other than bumping the versions of Go and golangci-lint in the workflow,
changes are pretty much just cosmetic. The one major fix is that the
golangci config file was named wrong and not respected, but it turns out
there wasn't much to fix.

One slight difference is that the pipeline now only checks the last two
minor versions of Go, rather than the last three. This is in alignment
with Go's official LTS policy. That said, I'm fine leaving 1.13 in the
go.mod, because there's no reason to actively remove support for older
Go versions if they work by chance. I just don't want to fail our
pipeline for them.

I'm opting not to upgrade go-cmp, because it's a test-only dependency
with no major changes that forces the minimum Go version up four years
into the future.

Signed-off-by: Rob Liebowitz <rliebowitz@morningconsult.com>